+++ pkgs/test/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * Annotate the classes used as annotations to restrict their usage to library
   level.
+* Stop required a `SILENT_OBSERVATORY` environment variable to run with
+  debugging and the JSON reporter.
 
 ## 1.16.0-nullsafety.4